feat: actualización del historial médico y reportería
